GOCBC-330: Fix panic during pipeline shutdown.
Motivation
----------
A race condition where a pipeline client was closed
before it fully finished initializing would cause
the SDK to panic trying to shut down its consumer.

Changes
-------
Rewrote a substantial portion of the memdPipelineClient
to attempt to simplify the execution flow.  We've also
significantly improved the inline comments throughout
the pipeline client to hopefully make it easier to
understand and debug.

Change-Id: I2e18f38ac026e60676d4adf7ab7683d146509fb0