GOCBC-333: Fail testing for abnormal logging.
Motivation
----------
Previously, errors and warnings that occurred during tests
could easily be lost among the noise of the test.  This
makes it harder to miss potential issues when running tests.

Changes
-------
Added a custom logger which keeps track of the number of
each level of log message that is emitted, and then added
a check to fail the tests if there are any abnormal logs
emitted.

Change-Id: Iec4d1f9b5794d6a2314899f3f239fd01961e53e5