GOCBC-331: Add internal shutdown notifications.
Motivation
----------
There has been discussion about the fact that after
calling the Close method on an Agent, there will still
be running goroutines left behind for a period of time.
This will ensure that everything is cleaned up in a
timely manner.

Changes
-------
Added an internal notification channel that is signalled
(closed) whenever the agent is closed.  There is additional
signalling back to the close method to indicate when those
additional goroutines have completed.

Change-Id: I029fe972d3d387e08384f40e418b6f348eaa2abb