Combine cfg bucket types and cccp based loopers
Motivation
----------
We should simplify the distinction between gcccp and cccp loopers
by combining them to reduce code duplication. We should also combine
cfgBucket and cfgCluster into one as cfgBucket is a superset of
cfgCluster anyway and there's loads of code duplication there too.

Changes
-------
Remove the gcccp looper and add a stop signal to the cccp looper.
Remove the cfgCluster and update BuildRouteConfig to support
a config that does not have a bucket.

Change-Id: I7c0930f230bc3d66a14092fb76e58c6c3832e40f