Update agent bootstrap to be implicit
Motivation
----------
In order to improve and simplify how we do initial agent bootstrap
(and initial cluster connect) we should push the addresses from
the connection stream straight into the muxers. The config pollers
can then use the config(/connections) from the muxers to try to
fetch the cluster config.

Changes
--------
Agent connect generates a route config containing kv server list
and  management server list using addresses from the connection
string.
Reinstate the poller controller which will try to start the cccp
looper. If that looper responds with a not found error (cccp
not supported) then fallback to http looper.
Config loopers will try to use the inital configurations to
try to fetch a config.
Expose a WaitUntilReady function that will call a callback when
a good config has been seen.

Change-Id: I1eee53ba5c49579f769d01d71688bcd1e971143d