Provide bootstrap list to config manager for network type heuristic
Motivation
----------
Instead of the providing the source server with every config to the
config manager we should provide all the bootstrap list nodes to
the manager on setup. We should then check the bootstrap list
against the first config to work out the network type. This removes
the implicit knowledge that source server on first iteration of
config pollers will be from the bootstrap list.

Changes
-------
No longer provide a source server alongside the config to the config
manager.
Pass the entire bootstrap list to the config manager on setup.
On first config use the whole bootstrap list (if required) to try
to work out the network type.

Change-Id: I3d6e5bfb902fd05b9beb06683132e7368a596025