Remove error parameter from PendingOp Cancel
Motivation
----------
PendingOp Cancel takes an error parameter which was previously
used so that gocb could indicate that a request was cancelled
due to a timeout. This error was propagated to circuit breakers.
Now that gocbcore handles timeouts directly we do not need
Cancel to take an error as all internal calls can call directly
into internalCancel.

Changes
-------
Remove the error parameter from Cancel.

Change-Id: I851babef8b31e0c3572e2d37a6de7c7bf18ec37b