Update DispatchDirectToAddress to take a pipeline directly
Motivation
----------
The way that we use DispatchDirectToAddress is slightly odd. In
all cases we pull the address off a pipeline and pass it in, we
then iterate pipelines until we find one that matches the address.
It'd be far easier (and prevent a possible race on client mux state
changing) if we just pass in a pipeline directly.

Changes
-------
Update DispectDirectToAddress to accept a pipeline as a parameter
in place of an address.

Change-Id: Ic9fd76b2884b3d9cfb9c3c9168019ad464400726