Correctly handle cancel for ping and http requests
Motivation
----------
Ping returns a PendingOp which can be cancelled. When that happens
we only cancel the kv ops. We need to also cancel HTTP ones.

Changes
-------
Add a Context field to httpRequest.
If Context is not nil then use that as the parent for the context
within the http component which is used for cancellation.
Provide a context from Ping which is only cancelled from the pingOp
Cancel function.

Change-Id: I8d4d2d1225182084f4a5839b3c7bceff57eae439