Don't try cccp polling if a non-default http port is used
Motivation
----------
If the user provides an address with a non-default http port
then we cannot connect to the cluster. We start the cccp poller
which sees no pipelines and loops back round to try again, and
again, and again... As the cccp looper cannot get a config and
does not fallback it means that we never setup any memd connections.

Changes
-------
If a non default http port is used then don't even attempt cccp
polling, skip straight to http polling.