GOCBC-853: Return empty certificate if authenticator returns nil
Motivation
----------
When the couchbases protocol is used (and therefore ssl enabled)
alongside PasswordAuthenticator and a root CA then we cause a panic
within the TLS library. This is because we always provide a
GetCertificate callback on our tls Config and if provided then the
library expects that to return a certificate or an error but we
return both as nil.

Changes
-------
Within the GetCertificate call, If the authenticator returns a nil
certificate and a nil error then actually return a blank certificate.

Change-Id: I9ea2066263c566aacbe7d46023a67f1271ddfc27