Move where request timer is stopped for kv reqs
Motivation
----------
When we receive a response from kv we stop the timer on the
corresponding request (assuming not cancelled). We do this too
early. If the request is retried then it's possible we can retry
forever because we'll never time it out.

Changes
-------
Move stopping the timer further down the processing until after
any retrying has been performed.

Change-Id: I41cc182a9f411b64f218c4ae4e058f8ea937bc4c